+++ src/avdecc/helper.cpp
@@ -538,6 +538,368 @@ QString audioClusterFormatToString(la::avdecc::entity::model::AudioClusterFormat
 	}
 }
 
+QString controlTypeToString(la::avdecc::entity::model::ControlType const controlType) noexcept
+{
+	switch (controlType)
+	{
+		case la::avdecc::entity::model::ControlType::Enable:
+			return "Enable";
+		case la::avdecc::entity::model::ControlType::Identify:
+			return "Identify";
+		case la::avdecc::entity::model::ControlType::Mute:
+			return "Mute";
+		case la::avdecc::entity::model::ControlType::Invert:
+			return "Invert";
+		case la::avdecc::entity::model::ControlType::Gain:
+			return "Gain";
+		case la::avdecc::entity::model::ControlType::Attenuate:
+			return "Attenuate";
+		case la::avdecc::entity::model::ControlType::Delay:
+			return "Delay";
+		case la::avdecc::entity::model::ControlType::SrcMode:
+			return "Sample Rate Converter Mode";
+		case la::avdecc::entity::model::ControlType::Snapshot:
+			return "Snapshot";
+		case la::avdecc::entity::model::ControlType::PowLineFreq:
+			return "Power Line Frequency";
+		case la::avdecc::entity::model::ControlType::PowerStatus:
+			return "Power Status";
+		case la::avdecc::entity::model::ControlType::FanStatus:
+			return "Fan Status";
+		case la::avdecc::entity::model::ControlType::Temperature:
+			return "Temperature";
+		case la::avdecc::entity::model::ControlType::Altitude:
+			return "Altitude";
+		case la::avdecc::entity::model::ControlType::AbsoluteHumidity:
+			return "Absolute Humidity";
+		case la::avdecc::entity::model::ControlType::RelativeHumidity:
+			return "Relative Humidity";
+		case la::avdecc::entity::model::ControlType::Orientation:
+			return "Orientation";
+		case la::avdecc::entity::model::ControlType::Velocity:
+			return "Velocity";
+		case la::avdecc::entity::model::ControlType::Acceleration:
+			return "Acceleration";
+		case la::avdecc::entity::model::ControlType::FilterResponse:
+			return "Filter Response";
+		case la::avdecc::entity::model::ControlType::Panpot:
+			return "Stereo Pan Position";
+		case la::avdecc::entity::model::ControlType::Phantom:
+			return "Phantom Power";
+		case la::avdecc::entity::model::ControlType::AudioScale:
+			return "Audio Scale";
+		case la::avdecc::entity::model::ControlType::AudioMeters:
+			return "Audio Meters";
+		case la::avdecc::entity::model::ControlType::AudioSpectrum:
+			return "Audio Spectrum";
+		case la::avdecc::entity::model::ControlType::ScanningMode:
+			return "Video Scanning Mode";
+		case la::avdecc::entity::model::ControlType::AutoExpMode:
+			return "Auto Exposure Mode";
+		case la::avdecc::entity::model::ControlType::AutoExpPrio:
+			return "Auto Exposure Priority";
+		case la::avdecc::entity::model::ControlType::ExpTime:
+			return "Exposure Time";
+		case la::avdecc::entity::model::ControlType::Focus:
+			return "Focus";
+		case la::avdecc::entity::model::ControlType::FocusAuto:
+			return "Focus Automatic";
+		case la::avdecc::entity::model::ControlType::Iris:
+			return "Iris";
+		case la::avdecc::entity::model::ControlType::Zoom:
+			return "Zoom";
+		case la::avdecc::entity::model::ControlType::Privacy:
+			return "Privacy";
+		case la::avdecc::entity::model::ControlType::Backlight:
+			return "Backlight Compensation";
+		case la::avdecc::entity::model::ControlType::Brightness:
+			return "Brightness";
+		case la::avdecc::entity::model::ControlType::Contrast:
+			return "Contrast";
+		case la::avdecc::entity::model::ControlType::Hue:
+			return "Hue";
+		case la::avdecc::entity::model::ControlType::Saturation:
+			return "Saturation";
+		case la::avdecc::entity::model::ControlType::Sharpness:
+			return "Sharpness";
+		case la::avdecc::entity::model::ControlType::Gamma:
+			return "Gamma";
+		case la::avdecc::entity::model::ControlType::WhiteBalTemp:
+			return "White Balance Temperature";
+		case la::avdecc::entity::model::ControlType::WhiteBalTempAuto:
+			return "White Balance Temperature Auto";
+		case la::avdecc::entity::model::ControlType::WhiteBalComp:
+			return "White Balance Components";
+		case la::avdecc::entity::model::ControlType::WhiteBalCompAuto:
+			return "White Balance Components Auto";
+		case la::avdecc::entity::model::ControlType::DigitalZoom:
+			return "Digital Zoom";
+		case la::avdecc::entity::model::ControlType::MediaPlaylist:
+			return "Media Playlist";
+		case la::avdecc::entity::model::ControlType::MediaPlaylistName:
+			return "Media Playlist Name";
+		case la::avdecc::entity::model::ControlType::MediaDisk:
+			return "Media Disk";
+		case la::avdecc::entity::model::ControlType::MediaDiskName:
+			return "Media Disk Name";
+		case la::avdecc::entity::model::ControlType::MediaTrack:
+			return "Media Track";
+		case la::avdecc::entity::model::ControlType::MediaTrackName:
+			return "Media Track Name";
+		case la::avdecc::entity::model::ControlType::MediaSpeed:
+			return "Media Speed";
+		case la::avdecc::entity::model::ControlType::MediaSamplePosition:
+			return "Media Sample Position";
+		case la::avdecc::entity::model::ControlType::MediaPlaybackTransport:
+			return "Media Playback Transport";
+		case la::avdecc::entity::model::ControlType::MediaRecordTransport:
+			return "Media Record Transport";
+		case la::avdecc::entity::model::ControlType::Frequency:
+			return "Frequency";
+		case la::avdecc::entity::model::ControlType::Modulation:
+			return "Modulation";
+		case la::avdecc::entity::model::ControlType::Polarization:
+			return "Polarization";
+		default:
+			AVDECC_ASSERT(false, "Not handled!");
+			return "Unknown";
+	}
+}
+
+QString controlValueTypeToString(la::avdecc::entity::model::ControlValueType::Type const controlValueType) noexcept
+{
+	switch (controlValueType)
+	{
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearInt8:
+			return "Linear Int 8";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearUInt8:
+			return "Linear UInt 8";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearInt16:
+			return "Linear Int 16";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearUInt16:
+			return "Linear UInt 16";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearInt32:
+			return "Linear Int 32";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearUInt32:
+			return "Linear UTnt 32";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearInt64:
+			return "Linear Int 64";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearUInt64:
+			return "Linear UInt 64";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearFloat:
+			return "Linear Float";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlLinearDouble:
+			return "Linear Double";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorInt8:
+			return "Selector Int 8";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorUInt8:
+			return "Selector UInt 8";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorInt16:
+			return "Selector Int 16";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorUInt16:
+			return "Selector UInt 16";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorInt32:
+			return "Selector Int 32";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorUInt32:
+			return "Selector UInt 32";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorInt64:
+			return "Selector Int 64";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorUInt64:
+			return "Selector UInt 64";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorFloat:
+			return "Selector Float";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorDouble:
+			return "Selector Double";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSelectorString:
+			return "Selector String";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayInt8:
+			return "Array Int 8";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayUInt8:
+			return "Array UInt 8";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayInt16:
+			return "Array Int 16";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayUInt16:
+			return "Array UInt 16";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayInt32:
+			return "Array Int 32";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayUInt32:
+			return "Array UInt 32";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayInt64:
+			return "Array Int 64";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayUInt64:
+			return "Array UInt 64";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayFloat:
+			return "Array Float";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlArrayDouble:
+			return "Array Double";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlUtf8:
+			return "UTF8";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlBodePlot:
+			return "Bode Plot";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSmpteTime:
+			return "SMPTE Time";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlSampleRate:
+			return "Sample Rate";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlGptpTime:
+			return "gPTP Time";
+		case la::avdecc::entity::model::ControlValueType::Type::ControlVendor:
+			return "Vendor";
+		case la::avdecc::entity::model::ControlValueType::Type::Expansion:
+			return "Expansion";
+		default:
+			AVDECC_ASSERT(false, "Not handled!");
+			return "Unknown";
+	}
+}
+
+QString controlValueUnitToString(la::avdecc::entity::model::ControlValueUnit::Unit const controlValueUnit) noexcept
+{
+	switch (controlValueUnit)
+	{
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Unitless:
+			return "Unitless";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Count:
+			return "Count";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Percent:
+			return "Percent";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::FStop:
+			return "fstop";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Seconds:
+			return "Seconds";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Minutes:
+			return "Minutes";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Hours:
+			return "Hours";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Days:
+			return "Days";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Months:
+			return "Months";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Years:
+			return "Years";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Samples:
+			return "Samples";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Frames:
+			return "Frames";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Hertz:
+			return "Hertz";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Semitones:
+			return "Semitones";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Cents:
+			return "Cents";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Octaves:
+			return "Octaves";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Fps:
+			return "FPS";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Metres:
+			return "Metres";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Kelvin:
+			return "Kelvin";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Grams:
+			return "Grams";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Volts:
+			return "Volts";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Dbv:
+			return "dBV";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Dbu:
+			return "dBu";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Amps:
+			return "Amps";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Watts:
+			return "Watts";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Dbm:
+			return "dBm";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Dbw:
+			return "dBW";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Pascals:
+			return "Pascals";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Bits:
+			return "Bits";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Bytes:
+			return "Bytes";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::KibiBytes:
+			return "KibiBytes";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::MebiBytes:
+			return "MebiBytes";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::GibiBytes:
+			return "GibiBytes";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::TebiBytes:
+			return "TebiBytes";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::BitsPerSec:
+			return "Bits Per Sec";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::BytesPerSec:
+			return "Bytes Per Sec";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::KibiBytesPerSec:
+			return "KibiBytes Per Sec";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::MebiBytesPerSec:
+			return "MebiBytes Per Sec";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::GibiBytesPerSec:
+			return "GibiBytes Per Sec";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::TebiBytesPerSec:
+			return "TebiBytes Per Sec";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Candelas:
+			return "Candelas";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Joules:
+			return "Joules";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Radians:
+			return "Radians";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Newtons:
+			return "Newtons";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Ohms:
+			return "Ohms";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::MetresPerSec:
+			return "Metres Per Sec";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::RadiansPerSec:
+			return "Radians Per Sec";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::MetresPerSecSquared:
+			return "Metres Per Sec Squared";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::RadiansPerSecSquared:
+			return "Radians Per Sec Squared";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Teslas:
+			return "Teslas";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Webers:
+			return "Webers";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::AmpsPerMetre:
+			return "Amps Per Metre";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::MetresSquared:
+			return "Metres Squared";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::MetresCubed:
+			return "Metres Cubed";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Litres:
+			return "Litres";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Db:
+			return "dB";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::DbPeak:
+			return "dB Peak";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::DbRms:
+			return "dB RMS";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Dbfs:
+			return "dBFS";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::DbfsPeak:
+			return "dBFS Peak";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::DbfsRms:
+			return "dBFS RMS";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Dbtp:
+			return "dBTP";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::DbSplA:
+			return "dB(A) SPL";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::DbZ:
+			return "dB(Z)";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::DbSplC:
+			return "dB(C) SPL";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::DbSpl:
+			return "dB SPL";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Lu:
+			return "LU";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::Lufs:
+			return "LUFS";
+		case la::avdecc::entity::model::ControlValueUnit::Unit::DbA:
+			return "dB(A)";
+		default:
+			AVDECC_ASSERT(false, "Not handled!");
+			return "Unknown";
+	}
+}
+
 QString memoryObjectTypeToString(la::avdecc::entity::model::MemoryObjectType const type) noexcept
 {
 	switch (type)